social_drf_api/requirements.txt
asgiref==3.8.1
black==25.1.0
click==8.2.1
django==5.2.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
PyJWT==2.9.0
psycopg[binary,pool]
pymongo==3.11.4  # Only keep if you're interacting with MongoDB directly (not via Djongo)
pytz==2025.2
setuptools==80.9.0
